+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 6.1.1(YYYY-MM-DD)
+
+### Enhancements 
+* None.
+
+### Fixed
+* [ObjectServer] Calling `SyncManager.refreshConnections()` did not correctly refresh connections in all cases, which could delay reconnects up to 5 minutes. (Issue [#7003](https://github.com/realm/realm-java/issues/7003))
+
+### Compatibility
+* Realm Object Server: 3.23.1 or later.
+* File format: Generates Realms with format v9 (Reads and upgrades all previous formats)
+* APIs are backwards compatible with all previous release of realm-java in the 6.x.y series.
+
+### Internal
+* None.
+
+
 ## 6.1.0(2020-01-17)
 
 ### Enhancements